requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
schedule>=1.2.0
